requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-dotenv
python-jose
fastapi
uvicorn